add placeholder images to Post List
