+++ sapper/src/post-components/PostNext.svelte
@@ -1,34 +1,14 @@
 <script>
-  import ExpandLink from '../layout-components/ExpandLink.svelte';
   import { Map } from 'immutable';
 
   export let post = Map();
   export let isPrevious = false;
 
+  import ExpandLink from '../layout-components/ExpandLink.svelte';
   import PostAvatar from './PostAvatar.svelte';
   import Image from '../document-components/Image.svelte';
   import H3 from '../document-components/H3.svelte';
-
-  const colorVars = [
-    '--filbert-lightGrey',
-    '--filbert-lightBlue',
-    '--filbert-lightOrange',
-    '--filbert-lightSalmon',
-  ];
-  function fisherYates(arg) {
-    const arr = [...arg];
-    let i = arr.length,
-      j,
-      temp;
-    while (--i > 0) {
-      j = Math.floor(Math.random() * (i + 1)); // Get random number ranging between 0 and i
-      temp = arr[j];
-      arr[j] = arr[i];
-      arr[i] = temp;
-    }
-    return arr;
-  }
-  const [toColor, fromColor] = fisherYates(colorVars);
+  import PostImagePlaceholder from './PostImagePlaceholder.svelte';
 </script>
 
 <style>
@@ -49,14 +29,6 @@
   }
   .post-avatar-container {
     padding: 4px;
-  }
-  .placeholder-image {
-    height: 400px;
-    border-radius: 4px;
-    background-image: radial-gradient(
-      var(--gradient-from-color),
-      var(--gradient-to-color)
-    );
   }
 
   @media (min-width: 768px) {
@@ -95,11 +67,7 @@
           viewPostOverride
         />
       {:else}
-        <div
-          class="placeholder-image"
-          style="--gradient-to-color: var({toColor}); --gradient-from-color:
-          var({fromColor});"
-        ></div>
+        <PostImagePlaceholder />
       {/if}
     </a>
   </div>
